requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
pgvector
langchain>=0.1.0
langchain-community>=0.1.0
openai
PyPDF2
python-multipart
requests
numpy
pypdf
tiktoken
langchain-openai
jinja2
email-validator
langchain-postgres